feat: create permission validation
