+++ src/commands/credentials.js
@@ -2,6 +2,7 @@ const { SlashCommandBuilder } = require('@discordjs/builders');
 const _ = require('lodash');
 const DiscordTools = require('../discordTools/discordTools.js');
 const { MessageEmbed } = require('discord.js');
+const { validatePermissions } = require('./base');
 
 module.exports = {
     data: new SlashCommandBuilder()
@@ -58,21 +59,9 @@ module.exports = {
     async execute(client, interaction) {
         let instance = client.readInstanceFile(interaction.guildId);
 
-        if (instance.role !== null) {
-            if (!interaction.member.permissions.has('ADMINISTRATOR') &&
-                !interaction.member.roles.cache.has(instance.role)) {
-                let role = DiscordTools.getRole(interaction.guildId, instance.role);
-                let str = `You are not part of the '${role.name}' role, therefore you can't run bot commands.`;
-                await client.interactionReply(interaction, {
-                    embeds: [new MessageEmbed()
-                        .setColor('#ff0040')
-                        .setDescription(`\`\`\`diff\n- ${str}\n\`\`\``)],
-                    ephemeral: true
-                });
-                client.log('WARNING', str);
-                return;
-            }
-        }
+		if(!validatePermissions(client, interaction, instance)) {
+			return;
+		}
 
         await interaction.deferReply({ ephemeral: true });
