+++ src/commands/market.js
@@ -1,6 +1,7 @@
 const { SlashCommandBuilder } = require('@discordjs/builders');
 const { MessageEmbed } = require('discord.js');
 const DiscordTools = require('../discordTools/discordTools');
+const { validatePermissions } = require('./base');
 
 module.exports = {
     data: new SlashCommandBuilder()
@@ -46,21 +47,9 @@ module.exports = {
     async execute(client, interaction) {
         let instance = client.readInstanceFile(interaction.guildId);
 
-        if (instance.role !== null) {
-            if (!interaction.member.permissions.has('ADMINISTRATOR') &&
-                !interaction.member.roles.cache.has(instance.role)) {
-                let role = DiscordTools.getRole(interaction.guildId, instance.role);
-                let str = `You are not part of the '${role.name}' role, therefore you can't run bot commands.`;
-                await client.interactionReply(interaction, {
-                    embeds: [new MessageEmbed()
-                        .setColor('#ff0040')
-                        .setDescription(`\`\`\`diff\n- ${str}\n\`\`\``)],
-                    ephemeral: true
-                });
-                client.log('WARNING', str);
-                return;
-            }
-        }
+		if(!validatePermissions(client, interaction, instance)) {
+			return;
+		}
 
         await interaction.deferReply({ ephemeral: true });
